Removing hearing prep utils
