+++ client/app/hearings/components/HearingWorksheetStream.jsx
@@ -5,10 +5,31 @@ import _ from 'lodash';
 import Button from '../../components/Button';
 import TabWindow from '../../components/TabWindow';
 import { onAddIssue } from '../actions/Issue';
-import { filterIssuesOnAppeal, currentIssues, priorIssues } from '../util/IssuesUtil';
 import HearingWorksheetIssues from './HearingWorksheetIssues';
 
 class HearingWorksheetStream extends Component {
+
+  filterIssuesOnAppeal = (issues, appealId) =>
+    _(issues).
+    omitBy('_destroy').
+    pickBy({ appeal_id: appealId }).
+    value();
+
+  currentIssues = (issues) => {
+    return _.omitBy(issues, (issue) => {
+      /* eslint-disable no-underscore-dangle */
+        return issue._destroy || (issue.disposition && !issue.disposition.includes('Remand') && issue.from_vacols);
+      /* eslint-enable no-underscore-dangle */
+    });
+};
+
+  priorIssues = (issues) => (
+    _.pickBy(issues, (issue) => (
+        /* eslint-disable no-underscore-dangle */
+        !issue._destroy && issue.disposition && !issue.disposition.includes('Remand') && issue.from_vacols
+        /* eslint-enable no-underscore-dangle */
+    ))
+);
 
   onAddIssue = (appealId) => () => this.props.onAddIssue(appealId, this.getVacolsSequenceId());
 
@@ -33,14 +54,14 @@ class HearingWorksheetStream extends Component {
     return <div> {Object.values(this.props.worksheetAppeals).map((appeal, key) => {
     /* eslint-enable array-callback-return */
 
-      const appealIssues = filterIssuesOnAppeal(this.props.worksheetIssues, appeal.id);
+      const appealIssues = this.filterIssuesOnAppeal(this.props.worksheetIssues, appeal.id);
 
       let appealWorksheetIssues;
 
       if (prior) {
-        appealWorksheetIssues = priorIssues(appealIssues);
+        appealWorksheetIssues = this.priorIssues(appealIssues);
       } else {
-        appealWorksheetIssues = currentIssues(appealIssues);
+        appealWorksheetIssues = this.currentIssues(appealIssues);
       }
 
       const currentIssueCount = issueCount;
@@ -73,9 +94,9 @@ class HearingWorksheetStream extends Component {
     </div>;
   };
 
-  getCurrentIssuesCount = () => _.size(currentIssues(this.props.worksheetIssues));
+  getCurrentIssuesCount = () => _.size(this.currentIssues(this.props.worksheetIssues));
 
-  getPriorIssuesCount = () => _.size(priorIssues(this.props.worksheetIssues));
+  getPriorIssuesCount = () => _.size(this.priorIssues(this.props.worksheetIssues));
 
   render() {
     const tabs = [{
